requirements.txt
aiosignal==1.3.1
appnope==0.1.3
argon2-cffi==21.1.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bleach==6.0.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==39.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.9.0
frozenlist==1.3.3
gmpy2==2.1.2
grpcio==1.53.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.15.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.2.1
msgpack==1.0.5
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==2.8.4
notebook==6.3.0
numexpr==2.8.4
numpy==1.23.4
opencv-python==4.6.0.66
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
ray==2.4.0
requests==2.29.0
scikit-learn==1.2.1
scipy==1.9.3
Send2Trash==1.8.0
setuptools==67.8.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
webencodings==0.5.1
wheel==0.38.4
widgetsnbextension==4.0.7
zipp==3.15.0
